requirements.txt
Flask
Flask-Cors
Flask-SQLAlchemy
Flask-Migrate

psycopg2-binary
GeoAlchemy2

Flask-SocketIO
eventlet

python-dotenv
gunicorn
